Cast memorder to int to fix traditional-converation from enum error

+++ include/fiber_atomic/atomic_gcc_clang.h
@@ -17,58 +17,58 @@ enum fiber_atomic_memorder {
 };
 
 #define fiber_atomic_memory_barrier(memorder) \
-	((void)__atomic_thread_fence(memorder))
+	((void)__atomic_thread_fence((int)memorder))
 
-#define fiber_atomic_load(ptr, memorder) (__atomic_load_n(ptr, memorder))
+#define fiber_atomic_load(ptr, memorder) (__atomic_load_n(ptr, (int)memorder))
 
 #define fiber_atomic_load_into(ptr, dest_ptr, memorder) \
-	((void)__atomic_load(ptr, dest_ptr, memorder))
+	((void)__atomic_load(ptr, dest_ptr, (int)memorder))
 
 #define fiber_atomic_store(ptr, val, memorder) \
-	((void)__atomic_store_n(ptr, val, memorder))
+	((void)__atomic_store_n(ptr, val, (int)memorder))
 
 #define fiber_atomic_store_from(ptr, src_ptr, memorder) \
-	((void)__atomic_store(ptr, src_ptr, memorder))
+	((void)__atomic_store(ptr, src_ptr, (int)memorder))
 
 #define fiber_atomic_exchange(ptr, val, memorder) \
-	(__atomic_exchange_n(ptr, val, memorder))
+	(__atomic_exchange_n(ptr, val, (int)memorder))
 
 #define fiber_atomic_exchange_into(ptr, src_ptr, res_ptr, memorder) \
-	((void)__atomic_exchange(ptr, src_ptr, res_ptr, memorder))
+	((void)__atomic_exchange(ptr, src_ptr, res_ptr, (int)memorder))
 
 #define fiber_atomic_cmp_xchng(ptr, expected_ptr, desired_val, is_weak,   \
 			       success_memorder, failure_memorder)        \
 	((int)__atomic_compare_exchange_n(ptr, expected_ptr, desired_val, \
-					  is_weak, success_memorder,      \
-					  failure_memorder))
+					  is_weak, (int)success_memorder, \
+					  (int)failure_memorder))
 
 #define fiber_atomic_cmp_xchng_from(ptr, expected_ptr, desired_ptr, is_weak, \
 				    success_memorder, failure_memorder)      \
 	((int)__atomic_compare_exchange(ptr, expected_ptr, desired_ptr,      \
-					is_weak, success_memorder,           \
-					failure_memorder))
+					is_weak, (int)success_memorder,      \
+					(int)failure_memorder))
 
 #define fiber_atomic_add_fetch(ptr, val, memorder) \
-	(__atomic_add_fetch(ptr, val, memorder))
+	(__atomic_add_fetch(ptr, val, (int)memorder))
 #define fiber_atomic_sub_fetch(ptr, val, memorder) \
-	(__atomic_sub_fetch(ptr, val, memorder))
+	(__atomic_sub_fetch(ptr, val, (int)memorder))
 #define fiber_atomic_and_fetch(ptr, val, memorder) \
-	(__atomic_and_fetch(ptr, val, memorder))
+	(__atomic_and_fetch(ptr, val, (int)memorder))
 #define fiber_atomic_xor_fetch(ptr, val, memorder) \
-	(__atomic_xor_fetch(ptr, val, memorder))
+	(__atomic_xor_fetch(ptr, val, (int)memorder))
 #define fiber_atomic_or_fetch(ptr, val, memorder) \
-	(__atomic_or_fetch(ptr, val, memorder))
+	(__atomic_or_fetch(ptr, val, (int)memorder))
 
 #define fiber_atomic_fetch_add(ptr, val, memorder) \
-	(__atomic_fetch_add(ptr, val, memorder))
+	(__atomic_fetch_add(ptr, val, (int)memorder))
 #define fiber_atomic_fetch_sub(ptr, val, memorder) \
-	(__atomic_fetch_sub(ptr, val, memorder))
+	(__atomic_fetch_sub(ptr, val, (int)memorder))
 #define fiber_atomic_fetch_and(ptr, val, memorder) \
-	(__atomic_fetch_and(ptr, val, memorder))
+	(__atomic_fetch_and(ptr, val, (int)memorder))
 #define fiber_atomic_fetch_xor(ptr, val, memorder) \
-	(__atomic_fetch_xor(ptr, val, memorder))
+	(__atomic_fetch_xor(ptr, val, (int)memorder))
 #define fiber_atomic_fetch_or(ptr, val, memorder) \
-	(__atomic_fetch_or(ptr, val, memorder))
+	(__atomic_fetch_or(ptr, val, (int)memorder))
 
 #define fiber_atomic_fetch_inc(ptr, memorder) \
 	fiber_atomic_fetch_add(ptr, 1, memorder)
